refactoring (parametrized utils for instrumentation)
